+++ src/environments/dsw.ts
@@ -8,9 +8,9 @@ export const dsw = {
     const: {
         themes: [
             {text: 'Default', file: ''},
-            {text: 'Contrast', file: 'themes/contrast.css'},
-            {text: 'Metro', file: 'themes/metro.css'},
-            {text: 'Black', file: 'themes/black.css'}
+            {text: 'Contrast', file: 'themes/contrast.css'}
+            // {text: 'Metro', file: 'themes/metro.css'},
+            // {text: 'Black', file: 'themes/black.css'}
         ],
         bgColorClasses: ['', 'cl1', 'cl2', 'cl3', 'cl4', 'cl5', 'cl6', 'cl7', 'cl8', 'cl9'],
         fontColors: ['fc1', 'fc2', 'fc3', 'fc4', 'fc5'],